feat: add VehicleController and VehicleService for vehicle management, including CRUD operations and dashboard summary
