+++ src/main/java/net/konic/vehicle/service/VehicleService.java
@@ -1,0 +1,71 @@
+package net.konic.vehicle.service;
+
+import net.konic.vehicle.dto.ApiResponse;
+import net.konic.vehicle.entity.Vehicle;
+import net.konic.vehicle.repository.VehicleRepository;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.stereotype.Service;
+import java.util.List;
+
+@Service
+public class VehicleService {
+
+    private final VehicleRepository vehicleRepository;
+
+    public VehicleService(VehicleRepository vehicleRepository) {
+        this.vehicleRepository = vehicleRepository;
+    }
+
+    // Create vehicle
+    @CacheEvict(value = {"vehicles", "vehicle"}, allEntries = true)
+    public Vehicle createVehicle(Vehicle vehicle) {
+        return vehicleRepository.save(vehicle);
+    }
+
+    // Get all vehicles
+    @Cacheable("vehicles")
+    public List<Vehicle> getAllVehicles() {
+        System.out.println("Fetching vehicles from DB...");
+        return vehicleRepository.findAll();
+    }
+
+    // Get vehicle by ID
+    @Cacheable(value = "vehicle", key = "#vehicleId")
+    public Vehicle getVehicleById(Long vehicleId) {
+        return vehicleRepository.findById(vehicleId)
+                .orElseThrow(() -> new RuntimeException("Vehicle not found: " + vehicleId));
+    }
+
+    // Update vehicle
+    @CacheEvict(value = {"vehicles", "vehicle"}, allEntries = true)
+    public Vehicle updateVehicle(Long vehicleId, Vehicle updatedVehicle) {
+        Vehicle existingVehicle = vehicleRepository.findById(vehicleId)
+                .orElseThrow(() -> new RuntimeException("Vehicle not found: " + vehicleId));
+
+        existingVehicle.setRegNumber(updatedVehicle.getRegNumber());
+        existingVehicle.setBrand(updatedVehicle.getBrand());
+        existingVehicle.setModel(updatedVehicle.getModel());
+        existingVehicle.setInsuranceExpiryDate(updatedVehicle.getInsuranceExpiryDate());
+        existingVehicle.setServiceDueDate(updatedVehicle.getServiceDueDate());
+        existingVehicle.setOwnerName(updatedVehicle.getOwnerName());
+        existingVehicle.setActive(updatedVehicle.isActive());
+
+        return vehicleRepository.save(existingVehicle);
+    }
+
+    // Delete vehicle
+    @CacheEvict(value = {"vehicles", "vehicle"}, allEntries = true)
+    public ApiResponse deleteVehicle(Long vehicleId) {
+        if (!vehicleRepository.existsById(vehicleId)) {
+            return new ApiResponse(false, "Vehicle not found: " + vehicleId);
+        }
+        vehicleRepository.deleteById(vehicleId);
+        return new ApiResponse(true, "Vehicle deleted successfully");
+    }
+
+    // Dashboard stats
+    public long getTotalVehicles() {
+        return vehicleRepository.count();
+    }
+}
